+++ src/main/java/com/hse/boards/repositories/BoardRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface BoardRepository extends CrudRepository<Board, Integer> {
+public interface BoardRepository extends CrudRepository<Board, Long> {
     @Query("SELECT b1.id, b1.name, b1.description FROM board_to_user b1 join board b2 on b1.boardId = b2.id where b1.userId = ?1")
-    Optional<List<Board>> findAllByUserId(Integer userId);
+    Optional<List<Board>> findAllByUserId(long userId);
 }
